Fix handling of SPH arrays and database changes between versions (#88)
* fix: Correctly read SPH data

Reshapes complete SPH data array to allow for correct indexing. To deal with old and new database versions, which may or may not include strain rate, flags for strain and strain rate were separated, as `isphfg(9)` can be 6 or 12.

* fix: Handle inconsistant SPH header between versions

There was an undocumented DB change between old and new versions where isphfg(1) can be 10 or 11. Data flag handling was modified to handle this behaviour.